pyproject.toml
[tool.poetry]
name = "pyunifi"
version = "0.1.0"
description = ""
authors = ["Bjørn Magnus Mathisen <bjornmagnus@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "*"
xlrd = "1.2.0"
openpyxl = "*"
python-dotenv = "*"
geopy = "*"
fuzzywuzzy = "*"
jupyter = "^1.0.0"
shapely = "^2.0.1"
geojson = "^3.0.1"
pyproj = "^3.6.0"
tqdm = "^4.66.1"
pymongo = "^4.5.0"
requests = "^2.31.0"
vcrpy = "^5.1.0"
confuse = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
flake8 = "^6.1.0"
debugpy = "^1.7.0"
rope = "^1.9.0"
python-lsp-server = {extras = ["all"], version = "^1.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
